Commit1
The UI, Main, Functionality, and CharacterObject classes have been created.

The UI consits of 2 Panels, each representing a character, and a Button to initiate an attack.

The character object method contains attributes for a name, health, compnent representing the character, an ability list, aswell as getters and setters.

Funvtionality consits of methods to:
-Generate a random number
-Delay logic based on user defined length
-Make a UI component flash
-Register hit logic

The main class just initialises objects and sets values.

Currently there is a bug, where the flash triggered by a button press causes the panel to display a broken image of a showmessage.